requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

joblib==1.5.1
scikit-learn==1.2.2
numpy==1.24.6
pandas==2.3.0
scipy==1.15.2

category-encoders==2.7.0
statsmodels==0.14.4
patsy==1.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
threadpoolctl==3.6.0
six==1.17.0
packaging==25.0

transformers 
torch 
torchvision 
pillow